♻️ Clean native includes and bounded parsing operations
Remove unused writer and binding includes and include each referenced type directly. Retain nanobind conversion headers with specific explanations. Parse trimmed FGL numbers through standard iterators, reserve known conversion-test input capacity, and make the stacked-layout size conversion explicit.

+++ bindings/mnt/pyfiction/src/pyfiction/layouts/cartesian_layout.cpp
@@ -135,7 +135,7 @@ void cartesian_layout(nanobind::module_& m, const char* name)
             [](const Lyt& lyt)
             {
                 std::vector<fiction::coordinate<Lyt>> coords{};
-                coords.reserve(lyt.area() * (lyt.z() + 1));
+                coords.reserve(lyt.area() * (static_cast<uint64_t>(lyt.z()) + 1u));
                 lyt.foreach_coordinate([&coords](const auto& c) { coords.push_back(c); });
                 return coords;
             },
